Fix: Enable clean compilation of mbedTLS build with -Wextra
Resolves all remaining compiler warnings flagged by '-Wextra' when
building with the mbedTLS crypto suite. This completes the primary
code hardening effort for all major build configurations.

The final set of fixes primarily addressed `-Wsign-compare` bugs in
the core libraries, test suites, and crypto abstraction layer, along
with several remaining `-Wmissing-field-initializers` and
`-Wunused-parameter` issues.

+++ ici/crypto/MBEDTLS_SUITES/secrypt.c
@@ -222,7 +222,7 @@ int crypt_and_hash_buffer(
     LARGE_INTEGER li_size;
     __int64 offset;
 #else
-    off_t offset = 0;
+    size_t  offset = 0;
 #endif
